Extract tests for equals to own file
